Fix code to match docs for the redirect URI
The README mentioned a `...URI` variable but the code used `...URL`.

This fixes the code to use `...URI` to match the documentation.

It also adds a note to the README to specify that the Spotify application dashboard must be configured with the callback URI so that Spotify will respect it when requested by the application.